+++ cmd/update.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/getantibody/antibody/antibody"
+	"github.com/getantibody/antibody/antibodylib"
 	"github.com/getantibody/antibody/project"
 	"github.com/spf13/cobra"
 )
@@ -12,8 +12,9 @@ var updateCmd = &cobra.Command{
 	Use:   "update",
 	Short: "updates all previously bundled plugins",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		fmt.Println("Updating all bundles in " + antibody.Home() + "...")
-		return project.Update(antibody.Home())
+		var home = antibodylib.Home()
+		fmt.Printf("Updating all bundles in %v...\n", home)
+		return project.Update(home)
 	},
 }
